# POC Fasten Mime4J with buffer recycling "a la" Jackson