--- a/core/src/main/java/org/apache/james/mime4j/io/MimeBoundaryInputStream.java
+++ b/core/src/main/java/org/apache/james/mime4j/io/MimeBoundaryInputStream.java
@@ -23,6 +23,7 @@ import org.apache.james.mime4j.MimeException;
 import org.apache.james.mime4j.MimeIOException;
 import org.apache.james.mime4j.util.ByteArrayBuffer;
 import org.apache.james.mime4j.util.CharsetUtil;
+import org.apache.james.mime4j.util.RecycledByteArrayBuffer;
 
 import java.io.IOException;
 
@@ -354,6 +355,11 @@ public class MimeBoundaryInputStream extends LineReaderInputStream {
 
     @Override
     public boolean unread(ByteArrayBuffer buf) {
+        return false;
+    }
+
+    @Override
+    public boolean unread(RecycledByteArrayBuffer buf) {
         return false;
     }
 }