… benchmark/pom.xml
… benchmark/src/main/java/org/apache/james/mime4j/JMHFieldBench.java
… benchmark/src/main/java/org/apache/james/mime4j/JMHLongMultipartReadBench.java
… benchmark/src/main/java/org/apache/james/mime4j/LongMultipartReadBench.java
… benchmark/src/main/resources/long-headers.msg
… benchmark/src/main/resources/long-multipart.msg
… core/pom.xml
… core/src/main/java/org/apache/james/mime4j/util/MimeParameterMapping.java
… core/src/main/java/org/apache/james/mime4j/util/RecycledByteArrayBuffer.java
… core/src/test/java/org/apache/james/mime4j/stream/DefaultFieldBuilderTest.java
… dom/pom.xml
… dom/src/main/java/org/apache/james/mime4j/dom/SingleBody.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentDescriptionField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentDispositionField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentIdField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentLanguageField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentLengthField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentLocationField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentMD5Field.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentTransferEncodingField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ContentTypeField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/MimeVersionField.java
… dom/src/main/java/org/apache/james/mime4j/dom/field/ParsedField.java
… dom/src/main/java/org/apache/james/mime4j/field/ContentDispositionFieldLenientImpl.java
… dom/src/main/java/org/apache/james/mime4j/field/Fields.java
… dom/src/main/java/org/apache/james/mime4j/message/BasicBodyFactory.java
… dom/src/main/java/org/apache/james/mime4j/message/DefaultBodyDescriptorBuilder.java
… dom/src/main/java/org/apache/james/mime4j/message/DefaultMessageBuilder.java
… dom/src/main/java/org/apache/james/mime4j/message/DefaultMessageWriter.java
… dom/src/test/java/org/apache/james/mime4j/field/FieldsTest.java