Inspired from what Jackson does...

# Before

```
Benchmark                          Mode  Cnt   Score   Error  Units
LongMultipartReadBench.benchmark1  avgt    5  29.898 ± 1.067  us/op
LongMultipartReadBench.benchmark2  avgt    5  30.392 ± 1.923  us/op
LongMultipartReadBench.benchmark3  avgt    5  35.179 ± 1.848  us/op
LongMultipartReadBench.benchmark4  avgt    5  58.478 ± 5.990  us/op

Alloc rate 4.5GB/s for the DOM API
```

# After

```
Benchmark                          Mode  Cnt   Score   Error  Units
LongMultipartReadBench.benchmark1  avgt    5  27.460 ± 5.143  us/op
LongMultipartReadBench.benchmark2  avgt    5  28.033 ± 1.699  us/op
LongMultipartReadBench.benchmark3  avgt    5  34.176 ± 9.447  us/op
LongMultipartReadBench.benchmark4  avgt    5  45.171 ± 2.774  us/op

Alloc rate: 3GB/s for the DOM API
```

We see a 25% improvment for the DOM API speed.

# How

 - Use commons-io ByteArrayOutputStream (expend is less costly)
 - Use buffer recycling where relevant
 - Contribute a JMH based benchmark allowing more reliable mesurements, and mesure memory churn

Debugging buffer recycling proved to be challenging, I resorted to sampling on missed allocation to identify spots not giving back their buffers, as well as double free disgnosing.

Segragating buffer recyclers per class proved to be really valuable to diagnose stuff as well prevent mis-usage in one class to contaminate the others.