@chibenwa 2022-06-21
I did also took the time to run the exact same benchmark on master and on the modified version of this code

## master
     JMHLongMultipartReadBench.benchmark4                               avgt    5      63.734 ±    4.293   us/op
     JMHLongMultipartReadBench.benchmark4:·gc.alloc.rate                avgt    5    4189.872 ±  285.825  MB/sec

## After

    JMHLongMultipartReadBench.benchmark4                               avgt   10     40.269 ±   0.916   us/op
    JMHLongMultipartReadBench.benchmark4:·gc.alloc.rate                avgt   10    617.446 ±  13.876  MB/sec