@chibenwa 2022-06-28
See https://github.com/apache/james-mime4j/pull/73 and https://github.com/apache/james-mime4j/pull/74